Correcting the logic to conditionally pull the acq. program.
Previously, the code was expecting the second column of the first
row to always contain the acquisition program information.
Unfortunately, this isn't always the case; many topics do not
contain this information. Furthermore, the markup for the site
does not consistently use `<tr>` tags as expected.

This commit pulls the acquisition program information in a manner
similar to the way the rest of the information is pulled; it finds
the `Acquisition Program:` header and pulls information from the
next field if any is present. This changes the results slightly.
Previously, topics like 'AF151-032' would have the ITAR notification
pulled as the acquisition program. Now the resulting topic object does
not have an `acquisition_program` property.